app/requirements.txt
google-cloud-bigquery
google-cloud-storage
Flask
tensorflow==2.16.1
Pillow==9.1.0
pandas
db-dtypes
tqdm
